requirements.txt
aniso8601==9.0.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.0
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyJWT==2.10.0
pytz==2024.2
six==1.16.0
Werkzeug==3.1.3
